refactor: inputbuffer 와 parser 분리
